+++ CHANGELOG.md
@@ -13,8 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Tooltip: refined zone story indication.
 * Tooltip: refined zone story, questline and campaign identification.
 * Tooltip: updated storyline quest counter to support Warband vs. individual progress.
+* Tooltip: updated counter and visual indication of in-progress and ready-for-turn-in storyline quests.
 * Tooltip: updated color theme for storyline quests (`green` - Warband completed quest, `dark green` - quest completed by current char + Warband, `yellow` - current available quest, `lighter yellow` - ongoing/in-progress quest, `white` - available not yet completed quest, `orange` - story/lore quest.
-* Tooltip: updated counter and visual indication of in-progress storyline quests.
 * Tooltip: quest titles in the Loremaster tooltip can now optionally be hidden.
 * Tooltip: adjusted quest tooltip width - tooltips with shorter content will now be stretched to fit the game's default tooltip.